requirements.txt
Flask==3.0.2
pika==1.3.2
psycopg2_binary==2.9.9
python-dotenv==1.0.1
